+++ src/components/AccountFilter/AccountFilter.test.tsx
@@ -1,4 +1,4 @@
-import { getFilteredAccounts, mapToFilteredArray, useAccountFilterUtils } from ".";
+import { mapToFilteredArray, useAccountFilterWithMapFilter } from ".";
 import { mockImplicitAccount } from "../../mocks/factories";
 import { fireEvent, render, screen } from "../../mocks/testUtils";
 import accountsSlice from "../../utils/store/accountsSlice";
@@ -15,7 +15,7 @@ afterEach(() => {
 });
 
 const Fixture: React.FC = () => {
-  const { filterElement } = useAccountFilterUtils();
+  const { filterElement } = useAccountFilterWithMapFilter();
 
   return filterElement;
 };
@@ -79,23 +79,5 @@ describe("AccountFilter", () => {
       "hello",
       "cool",
     ]);
-  });
-
-  test("getFilteredAccounts returns the right value", () => {
-    const accounts = [mockImplicitAccount(0), mockImplicitAccount(1), mockImplicitAccount(2)];
-
-    expect(getFilteredAccounts(accounts, [mockImplicitAccount(1).address])).toEqual([
-      mockImplicitAccount(1),
-    ]);
-    expect(getFilteredAccounts(accounts, [mockImplicitAccount(5).address])).toEqual([]);
-
-    expect(getFilteredAccounts(accounts, [])).toEqual(accounts);
-    expect(
-      getFilteredAccounts(accounts, [
-        mockImplicitAccount(0).address,
-        mockImplicitAccount(1).address,
-        mockImplicitAccount(2).address,
-      ])
-    ).toEqual(accounts);
   });
 });
